move vitessce tab into microscopy file
