+++ src/encoded/static/components/item-pages/FileView.js
@@ -13,7 +13,6 @@ import { store } from './../../store';
 
 import { SearchTableTitle } from './components/tables/ItemPageTable';
 import { ExperimentSetsTableTabView, EmbeddedExperimentSetSearchTable } from './components/tables/ExperimentSetTables';
-import { VitesscePlainContainer } from './components/Vitessce/VitesscePlainContainer';
 import { OverviewHeadingContainer } from './components/OverviewHeadingContainer';
 import { OverViewBodyItem, WrapInColumn } from './DefaultItemView';
 import WorkflowRunTracingView, { FileViewGraphSection } from './WorkflowRunTracingView';
@@ -60,8 +59,6 @@ export default class FileView extends WorkflowRunTracingView {
         const tabs = [];
 
         tabs.push(FileViewOverview.getTabObject(this.props, width));
-
-        tabs.push(FileViewVitessce.getTabObject(this.props, width));
 
         if (getItemType(context) === 'FileReference') {
             const expSetTableProps = {
@@ -379,171 +376,3 @@ export class RelatedFilesOverViewBlock extends React.PureComponent {
     }
 
 }
-
-
-function FileViewVitessce (props) {
-    const { context, windowWidth, width, schemas, href } = props;
-    const viewConfig = {
-        "coordinationSpace": {
-            "dataset": {
-                "A": "A"
-            },
-            "embeddingType": {
-                "A": "UMAP"
-            },
-            "spatialCellsLayer": {
-                "A": {
-                    "opacity": 1,
-                    "radius": 20,
-                    "stroked": false,
-                    "visible": true
-                }
-            }
-        },
-        "datasets": [
-            {
-                "files": [
-                    {
-                        "fileType": "anndata-cells.zarr",
-                        "options": {
-                            "factors": [
-                                "obs/marker_gene_0",
-                                "obs/marker_gene_1",
-                                "obs/marker_gene_2",
-                                "obs/marker_gene_3",
-                                "obs/marker_gene_4"
-                            ],
-                            "mappings": {
-                                "UMAP": {
-                                    "dims": [
-                                        0,
-                                        1
-                                    ],
-                                    "key": "obsm/X_umap"
-                                }
-                            },
-                            "xy": "obsm/X_spatial"
-                        },
-                        "type": "cells",
-                        "url": "https://assets.hubmapconsortium.org/ea4cfecb8495b36694d9a951510dc3c6/hubmap_ui/anndata-zarr/secondary_analysis.zarr"
-                    },
-                    {
-                        "fileType": "anndata-cell-sets.zarr",
-                        "options": [
-                            {
-                                "groupName": "Leiden",
-                                "setName": "obs/leiden"
-                            }
-                        ],
-                        "type": "cell-sets",
-                        "url": "https://assets.hubmapconsortium.org/ea4cfecb8495b36694d9a951510dc3c6/hubmap_ui/anndata-zarr/secondary_analysis.zarr"
-                    },
-                    {
-                        "fileType": "anndata-expression-matrix.zarr",
-                        "options": {
-                            "matrix": "X",
-                            "matrixGeneFilter": "var/marker_genes_for_heatmap"
-                        },
-                        "type": "expression-matrix",
-                        "url": "https://assets.hubmapconsortium.org/ea4cfecb8495b36694d9a951510dc3c6/hubmap_ui/anndata-zarr/secondary_analysis.zarr"
-                    }
-                ],
-                "name": "ea4cfecb8495b36694d9a951510dc3c6",
-                "uid": "A"
-            }
-        ],
-        "description": "",
-        "initStrategy": "auto",
-        "layout": [
-            {
-                "component": "scatterplot",
-                "coordinationScopes": {
-                    "dataset": "A",
-                    "embeddingType": "A"
-                },
-                "h": 6,
-                "w": 4,
-                "x": 0,
-                "y": 0
-            },
-            {
-                "component": "spatial",
-                "coordinationScopes": {
-                    "dataset": "A",
-                    "spatialCellsLayer": "A"
-                },
-                "h": 6,
-                "w": 5,
-                "x": 4,
-                "y": 0
-            },
-            {
-                "component": "cellSets",
-                "coordinationScopes": {
-                    "dataset": "A"
-                },
-                "h": 3,
-                "w": 3,
-                "x": 9,
-                "y": 0
-            },
-            {
-                "component": "genes",
-                "coordinationScopes": {
-                    "dataset": "A"
-                },
-                "h": 3,
-                "w": 3,
-                "x": 9,
-                "y": 4
-            },
-            {
-                "component": "heatmap",
-                "coordinationScopes": {
-                    "dataset": "A"
-                },
-                "h": 4,
-                "w": 7,
-                "x": 0,
-                "y": 6
-            },
-            {
-                "component": "cellSetExpression",
-                "coordinationScopes": {
-                    "dataset": "A"
-                },
-                "h": 4,
-                "w": 5,
-                "x": 7,
-                "y": 6
-            }
-        ],
-        "name": "ea4cfecb8495b36694d9a951510dc3c6",
-        "version": "1.0.4"
-    };
-    return (
-        <div>
-            <div className="row overview-blocks">
-                <VitesscePlainContainer height={800} width={1120} theme="light" viewConfig={viewConfig} />
-            </div>
-        </div>
-    );
-}
-FileViewVitessce.propTypes = {
-};
-FileViewVitessce.getTabObject = function({ context, schemas, windowWidth, href }, width){
-    return {
-        'tab' : <span><i className="icon icon-project-diagram fas icon-fw"/> Vitessce</span>,
-        'key' : 'file-vitessce',
-        //'disabled' : !Array.isArray(context.experiments),
-        'content' : (
-            <div className="overflow-hidden">
-                <h3 className="tab-section-title">
-                    <span>Vitessce Visualization</span>
-                </h3>
-                <hr className="tab-section-title-horiz-divider"/>
-                <FileViewVitessce {...{ context, width, windowWidth, schemas, href }} />
-            </div>
-        )
-    };
-};
